+++ ctx.h
@@ -5,7 +5,6 @@ public:
 					sleepTime(60) {
 		memset(mqttUser, 0, 128);
 		memset(mqttPassword, 0, 128);
-		memset(mqttFingerprint, 0, 128);
 		memset(mqttClientName, 0, 21);
 		strcpy(mqttUpdateStatusTopic, "stat/display/needUpdate");
 		strcpy(mqttCommandTopic, "cmd/display/upload");
@@ -22,9 +21,6 @@ public:
 		}
 		if (customMqttPassword) {
 			delete customMqttPassword;
-		}
-		if (customMqttFingerprint) {
-			delete customMqttFingerprint;
 		}
 		if (customMqttUpdateStatusTopic) {
 			delete customMqttUpdateStatusTopic;
@@ -45,7 +41,6 @@ public:
 		customMqttPort = new WiFiManagerParameter("port", "MQTT port", mqttPortAsString, 6);
 		customMqttUser = new WiFiManagerParameter("MQTT user", "MQTT user name", mqttUser, 128);
 		customMqttPassword = new WiFiManagerParameter("MQTT password", "MQTT password", mqttPassword, 128);
-		customMqttFingerprint = new WiFiManagerParameter("MQTT TLS fingerprint", "MQTT server certifacte fingerprint", mqttFingerprint, 128);
 		customMqttUpdateStatusTopic = new WiFiManagerParameter("updateTopic", "MQTT update topic", mqttUpdateStatusTopic, 128);
 		customMqttCommandTopic = new WiFiManagerParameter("commandTopic", "MQTT command topic", mqttCommandTopic, 128);
 		itoa(sleepTime, sleepTimeAsString, 10);
@@ -57,7 +52,6 @@ public:
 		mqttPort = atoi(customMqttPort->getValue());
 		strcpy(mqttUser, customMqttUser->getValue());
 		strcpy(mqttPassword, customMqttPassword->getValue());
-		strcpy(mqttFingerprint, customMqttFingerprint->getValue());
 		strcpy(mqttUpdateStatusTopic, customMqttUpdateStatusTopic->getValue());
 		strcpy(mqttCommandTopic, customMqttCommandTopic->getValue());
 		sleepTime = atoi(customSleepTime->getValue());
@@ -72,7 +66,6 @@ public:
 	int mqttPort;
 	char mqttUser[128];
 	char mqttPassword[128];
-	char mqttFingerprint[128];
 	char mqttClientName[21];
 	char mqttUpdateStatusTopic[128];
 	char mqttCommandTopic[128];
@@ -84,7 +77,6 @@ public:
 	WiFiManagerParameter *customMqttPort;
 	WiFiManagerParameter *customMqttUser;
 	WiFiManagerParameter *customMqttPassword;
-	WiFiManagerParameter *customMqttFingerprint;
 	WiFiManagerParameter *customMqttUpdateStatusTopic;
 	WiFiManagerParameter *customMqttCommandTopic;
 	WiFiManagerParameter *customSleepTime;